Implement 1D downsampling including LTTB algorithm
